change _federationB_base to federation (#3866) Signed-off-by: dylan-fan <289765648@qq.com>
